Refactor multi-camera tracking app around config-first runtime
Move the Python example to a small config-driven entrypoint, split runtime
helpers into python/utils, and replace the old monolithic unit test surface
with focused config, pipeline, worker, image, tracker, and main tests.

Simplify the example YAML so it reflects the real runtime contract:
input, inference, tracking, and output.optiview, with optional debug frame
saving via output.debug_dir and save_every=0 disabling saves by default.
