+++ examples/object-detection/multi-camera-people-detection-and-tracking/common/config.yaml
@@ -1,0 +1,35 @@
+model: assets/models/yolo_v8m_mpk.tar.gz  # Path to the detector model pack.
+
+streams:  # RTSP inputs, one entry per camera.
+  - rtsp://192.168.0.235:8554/src1
+  - rtsp://192.168.0.235:8554/src2
+  - rtsp://192.168.0.235:8554/src3
+  - rtsp://192.168.0.235:8554/src6
+  - rtsp://192.168.0.235:8554/src7
+  - rtsp://192.168.0.235:8554/src8
+
+input:  # RTSP input transport settings.
+  tcp: true        # Use TCP for RTSP transport.
+  latency_ms: 200  # RTSP receive latency in milliseconds.
+
+inference:  # Detector and runtime settings.
+  frames: 0                   # Frame limit per stream. 0 means unlimited.
+  fps: 0                      # Output video FPS. 0 means use source FPS.
+  bitrate_kbps: 2500          # H.264 bitrate for OptiView video output.
+  profile: true               # Print per-stream profiling information.
+  person_class_id: 0          # Class id to keep from detector output.
+  detection_threshold: null   # Optional score threshold override.
+  nms_iou_threshold: null     # Optional NMS IoU override.
+  top_k: null                 # Optional top-k override.
+
+tracking:  # Tracker settings.
+  iou_threshold: 0.3      # Minimum IoU to match an existing track.
+  max_missing_frames: 15  # Frames to keep a missing track alive.
+
+output:  # Live output plus optional debug saves.
+  optiview:  # Live video and JSON destination.
+    host: 192.168.0.235      # Host running the OptiView receiver/viewer.
+    video_port_base: 9000    # Base UDP port for video channels.
+    json_port_base: 9100     # Base UDP port for JSON channels.
+  debug_dir: null            # Optional debug overlay directory. Null disables saved frames.
+  save_every: 0              # Save every Nth debug frame. 0 disables saving.
