+++ examples/object-detection/multi-camera-people-detection-and-tracking/python/requirements.txt
@@ -1,2 +1,3 @@
 numpy
 opencv-python
+PyYAML
