+++ examples/object-detection/multi-camera-people-detection-and-tracking/python/tests/test_main.py
@@ -1,0 +1,38 @@
+"""Entrypoint tests for the multi-camera people tracking example."""
+
+from __future__ import annotations
+
+import subprocess
+import sys
+from pathlib import Path
+
+import pytest
+
+
+EXAMPLE_DIR = Path(__file__).resolve().parent.parent.parent
+MAIN_PY = EXAMPLE_DIR / "python" / "main.py"
+
+
+@pytest.mark.unit
+class TestMainEntrypoint:
+    def test_help_runs(self):
+        result = subprocess.run(
+            [sys.executable, str(MAIN_PY), "--help"],
+            capture_output=True,
+            text=True,
+            timeout=10,
+            cwd=str(EXAMPLE_DIR),
+        )
+        assert result.returncode == 0
+        assert "--config" in result.stdout
+
+    def test_missing_config_file_fails_cleanly(self):
+        result = subprocess.run(
+            [sys.executable, str(MAIN_PY), "--config", "does-not-exist.yaml"],
+            capture_output=True,
+            text=True,
+            timeout=10,
+            cwd=str(EXAMPLE_DIR),
+        )
+        assert result.returncode != 0
+        assert "config" in result.stderr.lower()
