+++ examples/object-detection/multi-camera-people-detection-and-tracking/python/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility modules for the Python multi-camera people tracking example."""
