+++ examples/object-detection/multi-camera-people-detection-and-tracking/python/utils/tracker.py
